refactor: 자동채 객체를 정의한 로직을 models 폴더로 이동한다
